requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
gunicorn==23.0.0
dj-database-url # Database URLs
psycopg2-binary # PostgreSQL database
whitenoise # Serving static files
django-cors-headers==4.4.0
djangorestframework-simplejwt
requests>=2.32.3
cloudinary